refactor: FeedbackPopup과 로직 코드를 :feature:settings 모듈로 옮김
