+++ feature/main/src/main/res/values/strings.xml
@@ -70,10 +70,4 @@
     <string name="meal_popup_close">닫기</string>
 
     <string name="schedule_popup_close">닫기</string>
-
-    <string name="feedback_popup_title">피드백 보내기</string>
-    <string name="feedback_popup_textfield_placeholder">블린더 사용 후기를 알려주세요!</string>
-    <string name="feedback_popup_textfield_error_empty">피드백을 입력해 주세요.</string>
-    <string name="feedback_popup_cancel_button">취소</string>
-    <string name="feedback_popup_send_button">전송</string>
 </resources>
